MAJOR BUGFIX
git-svn-id: https://silverstripe-ecommerce.googlecode.com/svn/modules/ecommerce_product_variation/trunk@3064 950b781c-c544-efab-2f99-90083ee69683

+++ code/dod/ProductWithVariationDecorator.php
@@ -433,7 +433,7 @@ class ProductWithVariationDecorator extends DataObjectDecorator {
 			}
 		}
 		if(count($arrayOfTypesToKeepForProduct)) {
-			DB::query("DELETE FROM \"Product_VariationAttributes\" WHERE \"ProductAttributeTypeID\" NOT IN (".implode(",", $arrayOfTypesToKeepForProduct).")");
+			DB::query("DELETE FROM \"Product_VariationAttributes\" WHERE \"ProductAttributeTypeID\" NOT IN (".implode(",", $arrayOfTypesToKeepForProduct).") AND ProductID = ".$this->owner->ID);
 		}
 	}
 }
